Re-organising the api module for better navigation.
